+++ src/KillerPdf.Engine/Authoring/PdfStrokeStyle.cs
@@ -1,0 +1,15 @@
+namespace KillerPdf.Engine.Authoring;
+
+public enum PdfLineCap
+{
+    Butt = 0,
+    Round = 1,
+    ProjectingSquare = 2
+}
+
+public enum PdfLineJoin
+{
+    Miter = 0,
+    Round = 1,
+    Bevel = 2
+}